+++ src/main/java/io/tofpu/speedbridge2/domain/misc/Score.java
@@ -2,18 +2,18 @@ package io.tofpu.speedbridge2.domain.misc;
 
 public final class Score implements Comparable<Score> {
     private final int scoredOn;
-    private final long score;
+    private final double score;
 
-    public static Score of(final int islandSlot, final long score) {
+    public static Score of(final int islandSlot, final double score) {
         return new Score(islandSlot, score);
     }
 
-    public Score(final int scoredOn, final long score) {
+    public Score(final int scoredOn, final double score) {
         this.scoredOn = scoredOn;
         this.score = score;
     }
 
-    public long getScore() {
+    public double getScore() {
         return score;
     }
